+++ Dockerfile
@@ -105,6 +105,9 @@ ENTRYPOINT ["./tests/Fixtures/Symfony/app/console"]
 CMD ["swoole:server:run"]
 
 FROM base as CliDev
+RUN apk add --no-cache git
+ENV COMPOSER_ALLOW_SUPERUSER="1"
+COPY --chown=app:runner --from=app-installer /usr/bin/composer /usr/local/bin/composer
 WORKDIR /usr/src/app/tests/Fixtures/Symfony/app
 ENTRYPOINT ["./console"]
 CMD ["swoole:server:run"]
