+++ src/Bridge/Symfony/Bundle/DependencyInjection/SwooleExtension.php
@@ -41,7 +41,6 @@ use Symfony\Component\DependencyInjection\Reference;
 use Symfony\Component\HttpKernel\DependencyInjection\Extension;
 use Symfony\Component\Messenger\MessageBusInterface;
 use Symfony\Component\Messenger\Transport\TransportFactoryInterface;
-use function ucfirst;
 
 final class SwooleExtension extends Extension
 {
